refactor: hooks to add liquidities
